+++ tensorflow_quantum/core/ops/parse_context.h
@@ -77,7 +77,7 @@ tensorflow::Status GetGradients(tensorflow::OpKernelContext* context,
 
 // Parses the number of samples from the 'num_samples' input tensor.
 tensorflow::Status GetNumSamples(tensorflow::OpKernelContext* context,
-                                 std::vector<uint64_t>* parsed_num_samples);
+                                 std::vector<unsigned int>* parsed_num_samples);
 
 }  // namespace tfq
